requirements.txt
numpy==1.16.1
httplib2==0.10.3
pandas==0.24.1
SQLAlchemy==1.2.5
pyodbc==4.0.27
elasticsearch==6.2.0
google_api_python_client==1.7.9
python_dateutil==2.8.0
mysql==0.0.2
mysql-connector-python==8.0.15
rpy2==2.9.3
pytz==2018.3
tzlocal==2.0.0
requests==2.9.1
